+++ Keyboard/ZYRemoteKeyboardView.h
@@ -1,7 +1,7 @@
 #import "headers.h"
 
 @interface CALayerHost : CALayer
-@property (nonatomic, assign) unsigned int contextId;
+@property (nonatomic, assign) NSUInteger contextId;
 @end
 
 @interface ZYRemoteKeyboardView : UIView {
@@ -9,5 +9,5 @@
 	NSString *_identifier;
 }
 @property (nonatomic, retain) CALayerHost *layerHost;
--(void) connectToKeyboardWindowForApp:(NSString*)identifier;
+- (void)connectToKeyboardWindowForApp:(NSString*)identifier;
 @end
